README: simplify explanation
I originally thought that this was a sub-par solution because I'm
re-calling datum() for the whole chart, but perhaps this is an accepted
pattern for D3 / NVD3 ( http://stackoverflow.com/a/26536685/156060 ). I
can't find counterexamples to this hypothesis, so that's something. :-)

If there were, say, an addDataPoint() or an addToSet() function, I'd
certainly try that out.

People.find().fetch() is all happening client-side, so that should be
plenty fast.